+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## [unreleased]
 
 
+ * Core: Added a new property *devModeNpmExecutionTimeoutSec* to the server config to increase the npm execution timeout
+   in dev mode on slow computers (default is 3min)
  * Sandbox App: The sandbox now uses the "real" message bus and therefore supports Apps that use the messageBus
    interceptor, and can also simulate communication with other Apps on the same page.
    Also, the Sandbox can now load itself properly ;-)